Add files via upload
EDA Report on a 500k sample of the main data set. I'll be working on this challenge and wanted to the share a data report as may help others get an in-depth overview of the data-set before diving into this challenge a bit more. Was created using the YData Profiling library in Python.
Will upload the notebook shortly.